requirements.txt
python-dotenv
transformers
tensorflow
langchain_core
langchain-community
langchain_openai
openai
streamlit